# rest-client/openapi-definition.json
{
    "components": {
        "schemas": {
...
            "ErrorDto": {
                "properties": {
                    "error": {
                        "description": "Error string",
                        "type": "string"
                    },
                    "message": {
                        "description": "Message",
                        "type": "string"
                    },
                    "statusCode": {
                        "description": "HTTP status code",
                        "type": "number"
                    }
                },
                "required": [
                    "statusCode",
                    "message",
                    "error"
                ],
                "type": "object"
            },
...
            "ProductL0Src": {
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "l0InstrumentLocation": {
                        "type": "string"
                    },
                    "metadataPersistentId": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "srs": {
                        "type": "string"
                    },
                    "uuid": {
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "uuid",
                    "name",
                    "srs",
                    "metadataPersistentId",
                    "l0InstrumentLocation"
                ],
                "type": "object"
            },
            "ProductL0SrcDto": {
                "properties": {
                    "l0InstrumentLocation": {
                        "description": "Location of instrument files",
                        "type": "string"
                    },
                    "metadataPersistentId": {
                        "description": "Persistent Id of final product ",
                        "type": "string"
                    },
                    "name": {
                        "description": "Name of product for display purposes - from gazeteer",
                        "type": "string"
                    },
                    "srs": {
                        "description": "Spatial Reference of product",
                        "type": "string"
                    },
                    "uuid": {
                        "description": "Unique identifier for reference purposes",
                        "type": "string"
                    }
                },
                "required": [
                    "uuid",
                    "name",
                    "srs",
                    "metadataPersistentId",
                    "l0InstrumentLocation"
                ],
                "type": "object"
            },
...
            "ProductL3Src": {
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "metadataPersistentId": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "productTifLocation": {
                        "type": "string"
                    },
                    "resolution": {
                        "type": "string"
                    },
                    "srs": {
                        "type": "string"
                    },
                    "uuid": {
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "uuid",
                    "name",
                    "srs",
                    "metadataPersistentId",
                    "resolution",
                    "productTifLocation"
                ],
                "type": "object"
            },
            "ProductL3SrcDto": {
                "properties": {
                    "metadataPersistentId": {
                        "description": "Persistent Id of final product ",
                        "type": "string"
                    },
                    "name": {
                        "description": "Name of product for display purposes - from gazeteer",
                        "type": "string"
                    },
                    "productTifLocation": {
                        "description": "Location of final product ",
                        "type": "string"
                    },
                    "resolution": {
                        "description": "Resolution of product for display purposes ",
                        "type": "string"
                    },
                    "srs": {
                        "description": "Spatial Reference of product",
                        "type": "string"
                    },
                    "uuid": {
                        "description": "Unique identifier for reference purposes",
                        "type": "string"
                    }
                },
                "required": [
                    "uuid",
                    "name",
                    "srs",
                    "metadataPersistentId",
                    "resolution",
                    "productTifLocation"
                ],
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "access-token": {
                "bearerFormat": "JWT",
                "scheme": "bearer",
                "type": "http"
            }
        }
    },
    "info": {
        "contact": {
            "email": "AusSeabed@ga.gov.au",
            "name": "AusSeabed",
            "url": "http://ausseabed.gov.au/"
        },
        "description": "The API description for the Ausseabed product catalogue inventory",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "title": "AusSeabed product catalogue",
        "version": "0.1.10"
    },
    "openapi": "3.0.0",
    "paths": {
...
        "/products/l0-src": {
            "get": {
                "operationId": "ProductsL0SrcController_findAll",
                "parameters": [],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/ProductL0Src"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": ""
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Unable to authenticate request."
                    },
                    "408": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Server took too long to respond."
                    }
                },
                "security": [
                    {
                        "access-token": []
                    }
                ],
                "tags": [
                    "products/l0-src"
                ]
            },
            "post": {
                "operationId": "ProductsL0SrcController_create",
                "parameters": [],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductL0SrcDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductL0Src"
                                }
                            }
                        },
                        "description": ""
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Unable to authenticate request."
                    },
                    "408": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Server took too long to respond."
                    }
                },
                "security": [
                    {
                        "access-token": []
                    }
                ],
                "tags": [
                    "products/l0-src"
                ]
            }
        },
        "/products/l0-src/{productId}": {
            "delete": {
                "operationId": "ProductsL0SrcController_delete",
                "parameters": [
                    {
                        "in": "path",
                        "name": "productId",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Could not find the survey"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Unable to authenticate request."
                    },
                    "408": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Server took too long to respond."
                    }
                },
                "security": [
                    {
                        "access-token": []
                    }
                ],
                "tags": [
                    "products/l0-src"
                ]
            },
            "get": {
                "operationId": "ProductsL0SrcController_findOne",
                "parameters": [
                    {
                        "in": "path",
                        "name": "productId",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductL0Src"
                                }
                            }
                        },
                        "description": ""
                    },
                    "400": {
                        "description": "Could not find the survey"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Unable to authenticate request."
                    },
                    "408": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Server took too long to respond."
                    }
                },
                "security": [
                    {
                        "access-token": []
                    }
                ],
                "tags": [
                    "products/l0-src"
                ]
            },
            "put": {
                "operationId": "ProductsL0SrcController_update",
                "parameters": [
                    {
                        "in": "path",
                        "name": "productId",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductL0SrcDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Could not find the survey"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Unable to authenticate request."
                    },
                    "408": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Server took too long to respond."
                    }
                },
                "security": [
                    {
                        "access-token": []
                    }
                ],
                "tags": [
                    "products/l0-src"
                ]
            }
        },
...
        "/products/l3-src": {
            "get": {
                "operationId": "ProductsL3SrcController_findAll",
                "parameters": [],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/ProductL3Src"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": ""
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Unable to authenticate request."
                    },
                    "408": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Server took too long to respond."
                    }
                },
                "security": [
                    {
                        "access-token": []
                    }
                ],
                "tags": [
                    "products/l3-src"
                ]
            },
            "post": {
                "operationId": "ProductsL3SrcController_create",
                "parameters": [],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductL3SrcDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductL3Src"
                                }
                            }
                        },
                        "description": ""
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Unable to authenticate request."
                    },
                    "408": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Server took too long to respond."
                    }
                },
                "security": [
                    {
                        "access-token": []
                    }
                ],
                "tags": [
                    "products/l3-src"
                ]
            }
        },
        "/products/l3-src/{productId}": {
            "delete": {
                "operationId": "ProductsL3SrcController_delete",
                "parameters": [
                    {
                        "in": "path",
                        "name": "productId",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Could not find the survey"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Unable to authenticate request."
                    },
                    "408": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Server took too long to respond."
                    }
                },
                "security": [
                    {
                        "access-token": []
                    }
                ],
                "tags": [
                    "products/l3-src"
                ]
            },
            "get": {
                "operationId": "ProductsL3SrcController_findOne",
                "parameters": [
                    {
                        "in": "path",
                        "name": "productId",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductL3Src"
                                }
                            }
                        },
                        "description": ""
                    },
                    "400": {
                        "description": "Could not find the survey"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Unable to authenticate request."
                    },
                    "408": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Server took too long to respond."
                    }
                },
                "security": [
                    {
                        "access-token": []
                    }
                ],
                "tags": [
                    "products/l3-src"
                ]
            },
            "put": {
                "operationId": "ProductsL3SrcController_update",
                "parameters": [
                    {
                        "in": "path",
                        "name": "productId",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductL3SrcDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Could not find the survey"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Unable to authenticate request."
                    },
                    "408": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Server took too long to respond."
                    }
                },
                "security": [
                    {
                        "access-token": []
                    }
                ],
                "tags": [
                    "products/l3-src"
                ]
            }
        },
...
    },
    "servers": [
        {
            "url": "/rest"
        }
    ],
...
}